feat: read only and max conn

+++ app/src/main/java/com/jhonju/ps3netsrv/server/commands/DeleteFileCommand.java
@@ -13,6 +13,10 @@ public class DeleteFileCommand extends FileCommand {
 
     @Override
     public void executeTask() throws PS3NetSrvException, IOException {
+        if (ctx.isReadOnly()) {
+            send(ERROR_CODE_BYTEARRAY);
+            throw new PS3NetSrvException("Failed to delete file: server is executing as read only");
+        }
         send(getFile().delete() ? SUCCESS_CODE_BYTEARRAY : ERROR_CODE_BYTEARRAY);
     }
 }
